requirements.txt
flowlauncher
pyspellchecker
nltk
pyperclip
wikipedia
warnings